requirements_.txt
annotated-types==0.6.0
argcomplete==1.10.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
compressed-rtf==1.0.6
confection==0.1.3
constantly==23.10.4
cryptography==41.0.5
cymem==2.0.8
daphne==3.0.2
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
dnspython==2.4.2
docx2txt==0.8
ebcdic==1.1.1
email-validator==2.1.0.post1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.0.0/en_core_web_md-2.0.0.tar.gz#sha256=ec77172fd76e21766c0aeaac2daf2c9d314a6f88f45d209d43f5bfa0ea2bafac
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
executing==2.0.1
extract-msg==0.28.7
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.4
IMAPClient==2.1.0
incremental==22.10.0
ipython==8.17.2
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
olefile==0.46
packaging==23.2
pandas==2.1.3
parso==0.8.3
pdfminer.six==20191110
pexpect==4.8.0
Pillow==10.1.0
preshed==3.0.9
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
six==1.12.0
skillNer==1.0.3
smart-open==6.4.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
textract==1.6.5
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
triton==2.1.0
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
xlrd==1.2.0
XlsxWriter==3.1.9
zope.interface==6.1
